fix: correct IdxSize cast and add missing Tuple import
- Fixed service.rs: changed limit cast back to u32 (IdxSize without bigidx feature)
- Fixed async_client.py: added missing Tuple import for type annotations
- All packages now compile cleanly with zero errors

+++ polaroid-grpc/src/service.rs
@@ -91,7 +91,7 @@ impl DataFrameService for PolaroidDataFrameService {
         // Apply row limit if specified
         if let Some(n_rows) = req.n_rows {
             if n_rows > 0 {
-                lf = lf.limit(n_rows as u64);
+                lf = lf.limit(n_rows as u32);
             }
         }
